+++ data/src/main/java/com/example/data/source/LocalNoteSource.kt
@@ -5,7 +5,7 @@ import com.example.data.entity.NoteModel
 interface LocalNoteSource {
 
     suspend fun obtainAll(): List<NoteModel>
-    suspend fun obtainById(id: Int): NoteModel
+    suspend fun obtainById(id: Long): NoteModel
     suspend fun create(entity: NoteModel)
     suspend fun update(entity: NoteModel)
     suspend fun delete(entity: NoteModel)
