+++ domain/src/main/java/com/example/domain/repository/NoteRepository.kt
@@ -5,7 +5,7 @@ import com.example.domain.model.Note
 interface NoteRepository {
 
     suspend fun getAllNotes(): List<Note>
-    suspend fun getNoteById(id: Int): Note
+    suspend fun getNoteById(id: Long): Note
     suspend fun createNewNote(note: Note)
     suspend fun updateNote(note: Note)
     suspend fun deleteNote(note: Note)
